refactor(cti)!: drop topology inference and infrastructure reasoning
The topology SDO described machines to stand up. Core CALDERA has nothing
that reads it, so it was a deployment blueprint with no consumer.

Deletes cti_topology_inference, cti_pipeline_stage4_topology,
cti_pipeline_stage3, cti_infra_aggregation, cti_refinement and
cti_knowledge_graph, plus the build_topology and refine_topology MCP
tools. Stage 3 emitted infrastructure hypotheses that only stage 4 read.

The pipeline is now stage 1 to stage 2: raw to IR to STIX. Entity
extraction stays, retargeted at CALDERA facts in the next commit.

BREAKING CHANGE: the x-cti-topology SDO, outputs_topology/, the
cti_pipeline_build_topology and cti_pipeline_refine_topology tools, and
the stage3-infra and stage4-topology pipeline steps are all removed.
