--- tests/test_cti_topology_enhancements.py
@@ -1,292 +1,0 @@
-import json
-
-import pytest
-
-
-def _rel(rid, rtype, source, target):
-    return {
-        "type": "relationship",
-        "id": f"relationship--00000000-0000-4000-8000-{rid:012d}",
-        "relationship_type": rtype,
-        "source_ref": source,
-        "target_ref": target,
-    }
-
-
-def test_topology_uses_infra_graph_network_services_software_and_targeted_users():
-    from plugins.mcp.app.utilities.cti_topology_inference import build_topology
-
-    actor = {
-        "type": "threat-actor",
-        "id": "threat-actor--00000000-0000-4000-8000-000000000001",
-        "name": "Actor",
-    }
-    finance = {
-        "type": "identity",
-        "id": "identity--00000000-0000-4000-8000-000000000002",
-        "name": "Finance",
-        "identity_class": "organization",
-    }
-    user = {
-        "type": "user-account",
-        "id": "user-account--00000000-0000-4000-8000-000000000003",
-        "user_id": "jdoe",
-        "display_name": "jdoe",
-    }
-    infra = {
-        "type": "infrastructure",
-        "id": "infrastructure--00000000-0000-4000-8000-000000000004",
-        "name": "File Server",
-        "infrastructure_types": ["staging"],
-        "x_cti_os": "windows",
-    }
-    traffic = {
-        "type": "network-traffic",
-        "id": "network-traffic--00000000-0000-4000-8000-000000000005",
-        "dst_port": 445,
-        "protocols": ["tcp"],
-    }
-    software = {
-        "type": "software",
-        "id": "software--00000000-0000-4000-8000-000000000006",
-        "name": "Windows Server",
-        "version": "2019",
-        "vendor": "Microsoft",
-        "cpe": "cpe:2.3:o:microsoft:windows_server_2019:1809:*:*:*:*:*:*:*",
-    }
-    ap = {
-        "type": "attack-pattern",
-        "id": "attack-pattern--00000000-0000-4000-8000-000000000007",
-        "name": "LSASS Memory",
-        "external_references": [
-            {"source_name": "mitre-attack", "external_id": "T1003.001"}
-        ],
-    }
-
-    bundle = {
-        "type": "bundle",
-        "id": "bundle--00000000-0000-4000-8000-000000000008",
-        "objects": [
-            actor, finance, user, infra, traffic, software, ap,
-            _rel(9, "targets", actor["id"], finance["id"]),
-            _rel(10, "uses", actor["id"], infra["id"]),
-            _rel(11, "hosts", infra["id"], traffic["id"]),
-            _rel(12, "hosts", infra["id"], software["id"]),
-            _rel(13, "located-at", user["id"], finance["id"]),
-            _rel(14, "uses", infra["id"], ap["id"]),
-        ],
-    }
-    taxonomy = {
-        "_raw_objects": [],
-        "attack_id_index": {
-            "T1003.001": {
-                "name": "LSASS Memory",
-                "x_mitre_platforms": ["Windows"],
-                "x_mitre_permissions_required": ["User"],
-                "kill_chain_phases": [
-                    {"kill_chain_name": "mitre-attack", "phase_name": "credential-access"}
-                ],
-            }
-        },
-    }
-
-    topology = build_topology(bundle, taxonomy)
-
-    assert topology["infrastructure_graph"]["reachable_infrastructure_refs"] == [infra["id"]]
-    host = topology["hosts"][0]
-    assert host["stix_id"] == infra["id"]
-    assert host["reachable_from_roots"] is True
-    assert "microsoft-ds" in host["services"]
-    assert host["network_services"][0]["port"] == 445
-    assert any(sw.get("cpe") == software["cpe"] for sw in host["software_required"])
-    assert topology["network_edges"][0]["service"] == "microsoft-ds"
-    assert topology["identities"][0]["targeted"] is True
-    assert topology["user_accounts"][0]["organization_unit"] == "Finance"
-    assert "Finance" in topology["user_accounts"][0]["ansible_role_input"]["user_tags"]
-    assert topology["attack_surface"]["required_permissions"] == ["User"]
-    assert "process" in topology["attack_surface"]["d3fend_artifacts"]
-
-
-def test_alphv_real_ae_plan_enrichment_materializes_plan_facts():
-    from plugins.mcp.app.cti_pipeline_stage4_topology import (
-        _enrich_topology_with_ae_plan,
-        _technique_ids_in_bundle,
-    )
-    from plugins.mcp.app.utilities.cti_ae_library_loader import (
-        ae_plan_to_stix,
-        discover_ae_plans,
-        find_plan_by_adversary,
-        parse_ae_plan,
-    )
-    from plugins.mcp.app.utilities.cti_topology_inference import build_topology
-
-    plan = find_plan_by_adversary(discover_ae_plans(), "alphv_blackcat")
-    if not plan:
-        pytest.skip("external AttackEvals AEL data is not installed")
-    assert "attackevals-ael" in plan.get("library", "")
-    ae_ir = parse_ae_plan(plan, taxonomy=None)
-    bundle = ae_plan_to_stix(ae_ir, taxonomy=None)
-
-    topology = build_topology(bundle, {})
-    enriched = _enrich_topology_with_ae_plan(
-        topology,
-        plan,
-        ae_ir,
-        _technique_ids_in_bundle(bundle),
-    )
-
-    expected_hosts = {
-        (h.get("hostname") or "").strip().lower()
-        for h in ae_ir.get("infrastructure") or []
-        if (h.get("hostname") or "").strip()
-    }
-    actual_hosts = {
-        (h.get("hostname") or h.get("name") or "").strip().lower().removeprefix("host-")
-        for h in enriched["hosts"]
-    }
-    assert expected_hosts <= actual_hosts
-    assert not any("synthesized" in h for h in actual_hosts)
-    assert any(h.get("role") == "dc" for h in enriched["hosts"])
-    assert any(h.get("role") == "backup" for h in enriched["hosts"])
-    assert set(enriched["emulation_plan"]["techniques"]) == {
-        (ap.get("id") or "").upper()
-        for ap in ae_ir.get("attack_patterns") or []
-        if ap.get("id")
-    }
-    assert set(enriched["emulation_plan"]["software"]) == {
-        (item.get("name") or "").strip().lower()
-        for field in ("software", "malware", "tools")
-        for item in ae_ir.get(field) or []
-        if isinstance(item, dict) and (item.get("name") or "").strip()
-    }
-    assert set(enriched["emulation_plan"]["users"]) == {
-        (u.get("username") or "").strip().lower()
-        for u in ae_ir.get("user_accounts") or []
-        if isinstance(u, dict) and (u.get("username") or "").strip()
-    }
-    assert set(ae_ir.get("file_paths") or []) <= set(enriched.get("file_paths") or [])
-    assert set(ae_ir.get("registry_keys") or []) <= set(enriched.get("registry_keys") or [])
-
-
-def test_fuse_bundles_merges_by_mitre_id_and_remaps_relationships():
-    from plugins.mcp.app.utilities.cti_fusion import fuse_bundles
-
-    ap1 = {
-        "type": "attack-pattern",
-        "id": "attack-pattern--00000000-0000-4000-8000-000000000001",
-        "name": "Technique A",
-        "external_references": [
-            {"source_name": "mitre-attack", "external_id": "T1234"}
-        ],
-    }
-    ap2 = {
-        "type": "attack-pattern",
-        "id": "attack-pattern--00000000-0000-4000-8000-000000000002",
-        "name": "Technique A Alias",
-        "description": "second source",
-        "external_references": [
-            {"source_name": "mitre-attack", "external_id": "T1234"}
-        ],
-    }
-    tool = {
-        "type": "tool",
-        "id": "tool--00000000-0000-4000-8000-000000000003",
-        "name": "Tool A",
-    }
-    rel = {
-        "type": "relationship",
-        "id": "relationship--00000000-0000-4000-8000-000000000004",
-        "relationship_type": "uses",
-        "source_ref": tool["id"],
-        "target_ref": ap2["id"],
-    }
-
-    fused = fuse_bundles([
-        {"type": "bundle", "id": "bundle--1", "objects": [ap1]},
-        {"type": "bundle", "id": "bundle--2", "objects": [ap2, tool, rel]},
-    ])
-
-    attack_patterns = [o for o in fused["objects"] if o.get("type") == "attack-pattern"]
-    relationships = [o for o in fused["objects"] if o.get("type") == "relationship"]
-    assert len(attack_patterns) == 1
-    assert attack_patterns[0]["description"] == "second source"
-    assert relationships[0]["target_ref"] == attack_patterns[0]["id"]
-
-
-def test_knowledge_graph_persists_to_sqlite(tmp_path):
-    from plugins.mcp.app.utilities.cti_knowledge_graph import persist_bundle_topology
-
-    infra = {
-        "type": "infrastructure",
-        "id": "infrastructure--00000000-0000-4000-8000-000000000001",
-        "name": "Host A",
-    }
-    bundle = {"type": "bundle", "id": "bundle--1", "objects": [infra]}
-    topology = {
-        "hosts": [{
-            "name": "host-host-a",
-            "stix_id": infra["id"],
-            "role": "server",
-            "platform": "linux",
-        }],
-        "network_edges": [],
-    }
-
-    result = persist_bundle_topology(bundle, topology, db_path=tmp_path / "kg.sqlite")
-    assert result["nodes_seen"] >= 2
-    assert (tmp_path / "kg.sqlite").is_file()
-
-
-def test_host_names_stay_anchored_to_their_infrastructure_sdo():
-    """derive_hosts and derive_networks build host names independently.
-
-    derive_networks rebuilds the name to map it back to an infrastructure
-    id. When the two drift the host is not dropped, it is silently moved
-    into "unanchored-net" with anchor_identity None, so the topology still
-    looks well formed. Assert on the anchoring, not on membership.
-    """
-    from plugins.mcp.app.utilities.cti_topology_inference import build_topology
-
-    infra = {
-        "type": "infrastructure",
-        "id": "infrastructure--aaaaaaaa-0000-4000-8000-000000000001",
-        "name": "Mail Relay",
-        "infrastructure_types": ["workstation"],
-        "x_cti_os": "windows",
-    }
-    identity = {
-        "type": "identity",
-        "id": "identity--bbbbbbbb-0000-4000-8000-000000000002",
-        "name": "Contoso",
-        "identity_class": "organization",
-    }
-    rel = {
-        "type": "relationship",
-        "id": "relationship--cccccccc-0000-4000-8000-000000000003",
-        "relationship_type": "located-at",
-        "source_ref": infra["id"],
-        "target_ref": identity["id"],
-    }
-    bundle = {"type": "bundle", "id": "bundle--1",
-              "objects": [infra, identity, rel]}
-
-    topology = build_topology(bundle, {})
-
-    assert topology["type"] == "x-cti-topology"
-    assert topology["id"].startswith("x-cti-topology--")
-
-    names = {h["name"] for h in topology["hosts"]}
-    assert names, "no hosts derived"
-    assert all(n.startswith("host-") for n in names), names
-
-    networks = topology.get("networks") or []
-    anchored = {m for net in networks if net.get("anchor_identity")
-                for m in net.get("members") or []}
-    unanchored = {m for net in networks if not net.get("anchor_identity")
-                  for m in net.get("members") or []}
-
-    assert names <= anchored, (
-        f"hosts lost their identity anchor: {sorted(names - anchored)}; "
-        f"unanchored={sorted(unanchored)}"
-    )
-    assert not unanchored, f"unexpected unanchored hosts: {sorted(unanchored)}"
